adjust python file to also change x and y instead of just flipping y
